vulkan.cpp: general refractoring

+++ app/src/main/cpp/winlator/vulkan.cpp
@@ -15,23 +15,25 @@
 
 PFN_vkGetInstanceProcAddr gip;
 
-const char *getNativeLibraryDir(JNIEnv *env, jobject context) {
-    char *native_libdir = NULL;
+const char *get_native_library_dir(JNIEnv *env, jobject context) {
+    char *native_libdir;
 
-    if (context != NULL) {
+    if (context != nullptr) {
         jclass class_ = env->FindClass("com/winlator/cmod/core/AppUtils");
         jmethodID getNativeLibraryDir = env->GetStaticMethodID(class_, "getNativeLibDir",
                                                                "(Landroid/content/Context;)Ljava/lang/String;");
         jstring nativeLibDir = static_cast<jstring>(env->CallStaticObjectMethod(class_,
                                                                                 getNativeLibraryDir,
                                                                                 context));
-        native_libdir = (char *)env->GetStringUTFChars(nativeLibDir, nullptr);
+        if (nativeLibDir)
+            native_libdir = (char *)env->GetStringUTFChars(nativeLibDir, nullptr);
     }
     return native_libdir;
 }
 
 const char *get_driver_path(JNIEnv *env, jobject context, const char *driver_name) {
     char *driver_path;
+    char *absolute_path;
 
     jclass contextWrapperClass = env->FindClass("android/content/ContextWrapper");
     jmethodID  getFilesDir = env->GetMethodID(contextWrapperClass, "getFilesDir", "()Ljava/io/File;");
@@ -40,10 +42,12 @@ const char *get_driver_path(JNIEnv *env, jobject context, const char *driver_nam
     jmethodID getAbsolutePath = env->GetMethodID(fileClass, "getAbsolutePath", "()Ljava/lang/String;");
     jstring absolutePath = static_cast<jstring>(env->CallObjectMethod(filesDirObj,
                                                                       getAbsolutePath));
-    const char *absolute_path = env->GetStringUTFChars(absolutePath, nullptr);
-    asprintf(&driver_path, "%s/contents/adrenotools/%s/", absolute_path, driver_name);
 
-    env->ReleaseStringUTFChars(absolutePath, absolute_path);
+    if (absolutePath) {
+        absolute_path = (char *)env->GetStringUTFChars(absolutePath, nullptr);
+        asprintf(&driver_path, "%s/contents/adrenotools/%s/", absolute_path, driver_name);
+        env->ReleaseStringUTFChars(absolutePath, absolute_path);
+    }
 
     return driver_path;
 }
@@ -60,38 +64,53 @@ const char *get_library_name(JNIEnv *env, jobject context, const char *driver_na
 
     jstring libraryName = static_cast<jstring>(env->CallObjectMethod(adrenotoolsManagerObj,getLibraryName, driverName));
 
-    library_name = (char *)env->GetStringUTFChars(libraryName, nullptr);
+    if (libraryName)
+        library_name = (char *)env->GetStringUTFChars(libraryName, nullptr);
 
     return library_name;
 }
 
-void *init_vulkan(JNIEnv  *env, jobject context, const char *driver_name, const char *nativeLibDir) {
-    if (!driver_name || !nativeLibDir || !strcmp(driver_name, "System"))
-        return dlopen("/system/lib64/libvulkan.so", RTLD_LAZY | RTLD_LOCAL);
-    else {
+void *init_original_vulkan() {
+    return dlopen("/system/lib64/libvulkan.so", RTLD_LOCAL | RTLD_NOW);
+}
+
+void *init_vulkan(JNIEnv  *env, jobject context, jstring driverName) {
         char *tmpdir;
+        void *vulkan_handle;
+
+        const char *driver_name = env->GetStringUTFChars(driverName, nullptr);
+
+        if (!strcmp(driver_name, "System"))
+            return init_original_vulkan();
 
         const char *driver_path = get_driver_path(env, context, driver_name);
         const char *library_name = get_library_name(env, context, driver_name);
+        const char *native_library_dir = get_native_library_dir(env, context);
 
-        asprintf(&tmpdir, "%s%s", driver_path, "temp");
+        if (driver_path) {
+            asprintf(&tmpdir, "%s%s", driver_path, "temp");
+            mkdir(tmpdir, S_IRWXU | S_IRWXG);
+        }
 
-        mkdir(tmpdir, S_IRWXU | S_IRWXG);
+        vulkan_handle = adrenotools_open_libvulkan(RTLD_LOCAL | RTLD_NOW, ADRENOTOOLS_DRIVER_CUSTOM, tmpdir, native_library_dir, driver_path, library_name, nullptr, nullptr);
 
-        return adrenotools_open_libvulkan(RTLD_NOW, ADRENOTOOLS_DRIVER_CUSTOM, tmpdir, nativeLibDir, driver_path, library_name, nullptr, nullptr);
-    }
+        if (!vulkan_handle)
+           return init_original_vulkan();
+
+        return vulkan_handle;
 }
 
-VkInstance create_instance(const char *driverName, JNIEnv *env, jobject context) {
+VkResult create_instance(jstring driverName, JNIEnv *env, jobject context, VkInstance *instance) {
     VkResult result;
-    VkInstance instance;
     VkInstanceCreateInfo create_info = {};
     void *vulkan_handle;
 
-    vulkan_handle = init_vulkan(env, context,driverName, getNativeLibraryDir(env, context));
+    if (driverName)
+        vulkan_handle = init_vulkan(env, context, driverName);
+    else
+        vulkan_handle = init_original_vulkan();
 
     gip = (PFN_vkGetInstanceProcAddr)dlsym(vulkan_handle, "vkGetInstanceProcAddr");
-
     PFN_vkCreateInstance createInstance = (PFN_vkCreateInstance)dlsym(vulkan_handle, "vkCreateInstance");
 
     create_info.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
@@ -101,21 +120,18 @@ VkInstance create_instance(const char *driverName, JNIEnv *env, jobject context)
     create_info.enabledLayerCount = 0;
     create_info.enabledExtensionCount = 0;
 
-    result = createInstance(&create_info, NULL, &instance);
+    result = createInstance(&create_info, NULL, instance);
 
-    if (result != VK_SUCCESS)
-        __android_log_print(ANDROID_LOG_DEBUG, "GPUInformation", "Failed to create instance: %d", result);
-
-    return instance;
-
+    return result;
 }
 
-std::vector<VkPhysicalDevice> get_physical_devices(VkInstance instance) {
+VkResult get_physical_devices(VkInstance instance, std::vector<VkPhysicalDevice> &physical_devices) {
     VkResult result = VK_ERROR_UNKNOWN;
-    std::vector<VkPhysicalDevice> physical_devices;
     uint32_t deviceCount;
 
     PFN_vkEnumeratePhysicalDevices enumeratePhysicalDevices = (PFN_vkEnumeratePhysicalDevices)gip(instance, "vkEnumeratePhysicalDevices");
+    if (!enumeratePhysicalDevices)
+        return VK_ERROR_INITIALIZATION_FAILED;
 
     enumeratePhysicalDevices(instance, &deviceCount, NULL);
     physical_devices.resize(deviceCount);
@@ -123,29 +139,35 @@ std::vector<VkPhysicalDevice> get_physical_devices(VkInstance instance) {
     if (deviceCount > 0)
         result = enumeratePhysicalDevices(instance, &deviceCount, physical_devices.data());
 
-    if (result != VK_SUCCESS)
-        __android_log_print(ANDROID_LOG_DEBUG, "GPUInformation", "Failed to enumerate devices: %d", result);
-
-    return physical_devices;
+    return result;
 }
 
 extern "C" JNIEXPORT jstring JNICALL
 Java_com_winlator_cmod_core_GPUInformation_getVersion(JNIEnv *env, jclass obj, jstring driverName, jobject context) {
     VkPhysicalDeviceProperties props = {};
+    std::vector<VkPhysicalDevice> pdevices;
     char *driverVersion;
-    char *driver_name;
     VkInstance instance;
 
-    if (driverName != NULL)
-        driver_name = (char *)env->GetStringUTFChars(driverName, nullptr);
-    else
-        driver_name = NULL;
+    if  (create_instance(driverName, env, context, &instance) != VK_SUCCESS) {
+        printf("Failed to create instance");
+        return env->NewStringUTF("Unknown");
+    }
 
-    instance = create_instance(driver_name, env, context);
+    if (get_physical_devices(instance, pdevices) != VK_SUCCESS) {
+        printf("Failed to query physical devices");
+        return env->NewStringUTF("Unknown");
+    }
+
     PFN_vkGetPhysicalDeviceProperties getPhysicalDeviceProperties = (PFN_vkGetPhysicalDeviceProperties)gip(instance, "vkGetPhysicalDeviceProperties");
     PFN_vkDestroyInstance destroyInstance = (PFN_vkDestroyInstance)gip(instance, "vkDestroyInstance");
 
-    for (const auto &pdevice: get_physical_devices(instance)) {
+    if (!getPhysicalDeviceProperties || !destroyInstance) {
+        printf("Failed to get function pointers");
+        return env->NewStringUTF("Unknown");
+    }
+
+    for (const auto &pdevice: pdevices) {
         getPhysicalDeviceProperties(pdevice, &props);
         uint32_t vk_driver_major = VK_VERSION_MAJOR(props.driverVersion);
         uint32_t vk_driver_minor = VK_VERSION_MINOR(props.driverVersion);
@@ -155,8 +177,6 @@ Java_com_winlator_cmod_core_GPUInformation_getVersion(JNIEnv *env, jclass obj, j
     }
 
     destroyInstance(instance, NULL);
-    if (driverName != NULL)
-        env->ReleaseStringUTFChars(driverName, driver_name);
 
     return (env->NewStringUTF(driverVersion));
 }
@@ -164,20 +184,29 @@ Java_com_winlator_cmod_core_GPUInformation_getVersion(JNIEnv *env, jclass obj, j
 extern "C" JNIEXPORT jstring JNICALL
 Java_com_winlator_cmod_core_GPUInformation_getVulkanVersion(JNIEnv *env, jclass obj, jstring driverName, jobject context) {
     VkPhysicalDeviceProperties props = {};
+    std::vector<VkPhysicalDevice> pdevices;
     char *vulkanVersion;
-    char *driver_name;
     VkInstance instance;
 
-    if (driverName != NULL)
-        driver_name = (char *)env->GetStringUTFChars(driverName, nullptr);
-    else
-        driver_name = NULL;
+    if  (create_instance(driverName, env, context, &instance) != VK_SUCCESS) {
+        printf("Failed to create instance");
+        return env->NewStringUTF("Unknown");
+    }
 
-    instance = create_instance(driver_name, env, context);
+    if (get_physical_devices(instance, pdevices) != VK_SUCCESS) {
+        printf("Failed to query physical devices");
+        return env->NewStringUTF("Unknown");
+    }
+
     PFN_vkGetPhysicalDeviceProperties getPhysicalDeviceProperties = (PFN_vkGetPhysicalDeviceProperties)gip(instance, "vkGetPhysicalDeviceProperties");
     PFN_vkDestroyInstance destroyInstance = (PFN_vkDestroyInstance)gip(instance, "vkDestroyInstance");
 
-    for (const auto &pdevice: get_physical_devices(instance)) {
+    if (!getPhysicalDeviceProperties || !destroyInstance) {
+        printf("Failed to get function pointers");
+        return env->NewStringUTF("Unknown");
+    }
+
+    for (const auto &pdevice: pdevices) {
         getPhysicalDeviceProperties(pdevice, &props);
         uint32_t vk_driver_major = VK_VERSION_MAJOR(props.apiVersion);
         uint32_t vk_driver_minor = VK_VERSION_MINOR(props.apiVersion);
@@ -187,8 +216,6 @@ Java_com_winlator_cmod_core_GPUInformation_getVulkanVersion(JNIEnv *env, jclass 
     }
 
     destroyInstance(instance, NULL);
-    if (driverName != NULL)
-        env->ReleaseStringUTFChars(driverName, driver_name);
 
     return (env->NewStringUTF(vulkanVersion));
 }
@@ -196,78 +223,66 @@ Java_com_winlator_cmod_core_GPUInformation_getVulkanVersion(JNIEnv *env, jclass 
 extern "C" JNIEXPORT jstring JNICALL
 Java_com_winlator_cmod_core_GPUInformation_getRenderer(JNIEnv *env, jclass obj, jstring driverName, jobject context) {
     VkPhysicalDeviceProperties props = {};
+    std::vector<VkPhysicalDevice> pdevices;
     char *renderer;
-    char *driver_name;
     VkInstance instance;
 
-    if (driverName != NULL)
-        driver_name = (char *)env->GetStringUTFChars(driverName, nullptr);
-    else
-        driver_name = NULL;
+    if  (create_instance(driverName, env, context, &instance) != VK_SUCCESS) {
+        printf("Failed to create instance");
+        return env->NewStringUTF("Unknown");
+    }
 
-    instance = create_instance(driver_name, env, context);
+    if (get_physical_devices(instance, pdevices) != VK_SUCCESS) {
+        printf("Failed to query physical devices");
+        return env->NewStringUTF("Unknown");
+    }
+
     PFN_vkGetPhysicalDeviceProperties getPhysicalDeviceProperties = (PFN_vkGetPhysicalDeviceProperties)gip(instance, "vkGetPhysicalDeviceProperties");
     PFN_vkDestroyInstance destroyInstance = (PFN_vkDestroyInstance)gip(instance, "vkDestroyInstance");
 
-    for (const auto &pdevice: get_physical_devices(instance)) {
+    if (!getPhysicalDeviceProperties || !destroyInstance) {
+        printf("Failed to get function pointers");
+        return env->NewStringUTF("Unknown");
+    }
+
+    for (const auto &pdevice: pdevices) {
         getPhysicalDeviceProperties(pdevice, &props);
         asprintf(&renderer, "%s", props.deviceName);
     }
 
     destroyInstance(instance, NULL);
-    if (driverName != NULL)
-        env->ReleaseStringUTFChars(driverName, driver_name);
 
     return (env->NewStringUTF(renderer));
-}
-
-extern "C" JNIEXPORT jlong JNICALL
-Java_com_winlator_cmod_core_GPUInformation_getMemorySize(JNIEnv *env, jclass obj, jstring driverName, jobject context) {
-    VkPhysicalDeviceMemoryProperties props = {};
-    long memorySize;
-    char *driver_name;
-    VkInstance instance;
-
-    if (driverName != NULL)
-        driver_name = (char *)env->GetStringUTFChars(driverName, nullptr);
-    else
-        driver_name = NULL;
-
-    instance = create_instance(driver_name, env, context);
-    PFN_vkGetPhysicalDeviceMemoryProperties getPhysicalDeviceMemoryProperties = (PFN_vkGetPhysicalDeviceMemoryProperties)gip(instance, "vkGetPhysicalDeviceMemoryProperties");
-    PFN_vkDestroyInstance destroyInstance = (PFN_vkDestroyInstance)gip(instance, "vkDestroyInstance");
-
-    for (const auto &pdevice : get_physical_devices(instance)) {
-        getPhysicalDeviceMemoryProperties(pdevice, &props);
-        memorySize = props.memoryHeaps[0].size;
-    }
-
-    destroyInstance(instance, NULL);
-    if (driverName != NULL)
-        env->ReleaseStringUTFChars(driverName, driver_name);
-
-    return memorySize / 1048576;
 }
 
 extern "C" JNIEXPORT jobjectArray JNICALL
 Java_com_winlator_cmod_core_GPUInformation_enumerateExtensions(JNIEnv *env, jclass obj, jstring driverName, jobject context) {
     jobjectArray extensions;
     VkInstance instance;
+    std::vector<VkPhysicalDevice> pdevices;
     uint32_t extensionCount;
-    char *driver_name;
     std::vector<VkExtensionProperties> extensionProperties;
 
-    if (driverName != NULL)
-        driver_name = (char *)env->GetStringUTFChars(driverName, nullptr);
-    else
-        driver_name = NULL;
+    if  (create_instance(driverName, env, context, &instance) != VK_SUCCESS) {
+        printf("Failed to create instance");
+        return env->NewObjectArray(0, env->FindClass("java/lang/String"), env->NewStringUTF(""));
+    }
 
-    instance = create_instance(driver_name, env, context);
+    if (get_physical_devices(instance, pdevices) != VK_SUCCESS) {
+        printf("Failed to query physical devices");
+        return env->NewObjectArray(0, env->FindClass("java/lang/String"), env->NewStringUTF(""));
+    }
 
     PFN_vkEnumerateDeviceExtensionProperties enumerateDeviceExtensionProperties = (PFN_vkEnumerateDeviceExtensionProperties)gip(instance, "vkEnumerateDeviceExtensionProperties");
     PFN_vkDestroyInstance destroyInstance = (PFN_vkDestroyInstance)gip(instance, "vkDestroyInstance");
 
-    for (const auto &pdevice : get_physical_devices(instance)) {
+    if (!enumerateDeviceExtensionProperties || !destroyInstance) {
+        printf("Failed to get function pointers");
+        return env->NewObjectArray(0, env->FindClass("java/lang/String"), env->NewStringUTF(""));
+    }
+
+
+    for (const auto &pdevice : pdevices) {
         enumerateDeviceExtensionProperties(pdevice, NULL, &extensionCount, NULL);
         extensionProperties.resize(extensionCount);
         enumerateDeviceExtensionProperties(pdevice, NULL, &extensionCount, extensionProperties.data());
@@ -280,8 +295,6 @@ Java_com_winlator_cmod_core_GPUInformation_enumerateExtensions(JNIEnv *env, jcla
     }
 
     destroyInstance(instance, NULL);
-    if (driverName != NULL)
-        env->ReleaseStringUTFChars(driverName, driver_name);
 
     return extensions;
 }
